+++ src/main.ts
@@ -11,6 +11,8 @@ import { TrimWhitespaceSettingTab } from "./settings";
 import handleTextTrim from "./utils/trimText";
 
 const DEFAULT_SETTINGS: TrimWhitespaceSettings = {
+	TrimOnSave: true,
+
 	AutoTrimDocument: true,
 	AutoTrimTimeout: 2.5,
 
@@ -78,7 +80,7 @@ export default class TrimWhitespace extends Plugin {
 
 		if (typeof save === "function") {
 			saveCommandDefinition.callback = () => {
-				if (this.settings.AutoTrimDocument) {
+				if (this.settings.TrimOnSave) {
 					this.trimDocument();
 				}