+++ changelog.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - Plugin not loading on iOS (thanks @velebit!) ([\#14](https://github.com/zlovatt/obsidian-trim-whitespace/issues/14))
+- Typo in settings ([\#16](https://github.com/zlovatt/obsidian-trim-whitespace/issues/16))
 
 ---