+++ changelog.md
@@ -70,7 +70,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - Initial commit!
 
-[Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/master...develop
+[Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/main...develop
 [v0.2.2]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.1.0...v0.2.0