+++ changelog.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - Trimming via `CTRL/CMD + S` no longer requires Auto-Trim to be enabled ([\#12](https://github.com/zlovatt/obsidian-trim-whitespace/issues/12))
 - "Skip Code Blocks" setting name to "Preserve Code Blocks"
+- Improved detection of code blocks
 
 ### Fixed