+++ changelog.md
@@ -10,6 +10,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+-
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+---
+
+## [v0.3.0] - 2023/07/30
+
+### Added
+
 - Jest test suite (thanks @velebit!)
 - Setting to preserve leading characters if part of an indented list
 - Setting to control whether `CTRL/CMD + S` trims document
@@ -76,6 +92,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Initial commit!
 
 [Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/main...develop
+[v0.3.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.2...v0.3.0
 [v0.2.2]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.1.0...v0.2.0